+++ components/collector/src/source_collectors/gitlab/json_types.py
@@ -26,8 +26,13 @@ class GitLabJSON:
 class Pipeline(GitLabJSON):
     """GitLab pipeline JSON. See https://docs.gitlab.com/ee/api/pipelines.html.
 
-    We don't use the pipeline details to determine the pipeline duration, but simply subtract the update_at and
-    created_at timestamps. Reason is that GitLab itself is pretty unclear about what exactly the pipeline duration is.
+    To determine the bruto duration of a pipeline, we don't get the duration reported by GitLab, but simply subtract
+    the updated_at and created_at timestamps. Reason is that the API to list pipelines of a project does not return
+    the duration.Getting the duration requires  a call to the API to get one pipelines for each pipeline.
+
+    Note that GitLab does not report a duration for pipelines that have not finished.
+
+    Also note that GitLab itself is pretty unclear about what exactly the pipeline duration is.
     See https://gitlab.com/gitlab-org/gitlab/-/issues/19594.
     """
 
@@ -39,7 +44,7 @@ class Pipeline(GitLabJSON):
     source: str
     created_at: str
     web_url: str
-    duration: int = 0
+    duration: int | None = None
     updated_at: str = ""
     schedule_description: str = ""  # Pipeline schedule description for scheduled pipelines
 
@@ -58,11 +63,7 @@ class Pipeline(GitLabJSON):
     @property
     def netto_duration(self) -> timedelta:
         """Return the netto duration of the pipeline, meaning the duration excluding idle time."""
-        return timedelta(minutes=self.duration)
-
-    def pipeline_duration(self, exclude_idle_time: bool) -> timedelta:
-        """Return the duration of the pipeline."""
-        return self.netto_duration if exclude_idle_time else self.bruto_duration
+        return timedelta(seconds=self.duration or 0)
 
 
 @dataclass